Refactor(auth): 공통응답 및 공통예외처리 적용
