+++ src/main/java/com/example/deliveryappproject/domain/auth/dto/response/AuthAccessResponse.java
@@ -1,0 +1,11 @@
+package com.example.deliveryappproject.domain.auth.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class AuthAccessResponse {
+
+    private final String accessToken;
+}
