+++ src/main/java/com/example/deliveryappproject/web/auth/controller/AuthController.java
@@ -3,7 +3,9 @@ package com.example.deliveryappproject.web.auth.controller;
 import com.example.deliveryappproject.common.annotation.Auth;
 import com.example.deliveryappproject.common.annotation.RefreshToken;
 import com.example.deliveryappproject.common.dto.AuthUser;
+import com.example.deliveryappproject.common.response.Response;
 import com.example.deliveryappproject.domain.auth.dto.request.AuthLoginRequest;
+import com.example.deliveryappproject.domain.auth.dto.response.AuthAccessResponse;
 import com.example.deliveryappproject.domain.auth.dto.response.AuthTokenResponse;
 import com.example.deliveryappproject.domain.auth.service.AuthService;
 import jakarta.servlet.http.Cookie;
@@ -22,21 +24,27 @@ public class AuthController {
     private final AuthService authService;
 
     @PostMapping("/login")
-    public AuthTokenResponse login(@Valid @RequestBody AuthLoginRequest authLoginRequest, HttpServletResponse response) {
+    public Response<AuthAccessResponse> login(@Valid @RequestBody AuthLoginRequest authLoginRequest, HttpServletResponse response) {
         AuthTokenResponse authTokenResponse = authService.login(authLoginRequest);
+
         setRefreshTokenCookie(response, authTokenResponse.getRefreshToken());
-        return authTokenResponse;
+        AuthAccessResponse authAccessResponse = new AuthAccessResponse(authTokenResponse.getAccessToken());
+        return Response.of(authAccessResponse);
     }
 
     @GetMapping("/logout")
-    public void logout(@Auth AuthUser authUser) {
+    public Response<Void> logout(@Auth AuthUser authUser) {
         authService.logout(authUser);
+        return Response.empty();
     }
 
     @GetMapping("/refresh")
-    public void reissueAccessToken(@RefreshToken String refreshToken, HttpServletResponse response) {
+    public Response<AuthAccessResponse> reissueAccessToken(@RefreshToken String refreshToken, HttpServletResponse response) {
         AuthTokenResponse authTokenResponse = authService.reissueAccessToken(refreshToken);
+
         setRefreshTokenCookie(response, authTokenResponse.getRefreshToken());
+        AuthAccessResponse authAccessResponse = new AuthAccessResponse(authTokenResponse.getAccessToken());
+        return Response.of(authAccessResponse);
     }
 
     private void setRefreshTokenCookie(HttpServletResponse response, String refreshToken) {
