+++ src/test/java/com/example/deliveryappproject/domain/notice/service/NoticeServiceTest.java
@@ -6,7 +6,7 @@ import com.example.deliveryappproject.domain.notice.repository.NoticeRepository;
 import com.example.deliveryappproject.domain.store.entity.Store;
 import com.example.deliveryappproject.domain.store.repository.StoreRepository;
 import com.example.deliveryappproject.domain.user.entity.User;
-import com.example.deliveryappproject.domain.user.entity.UserRole;
+import com.example.deliveryappproject.domain.user.enums.UserRole;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
