+++ src/test/java/com/example/deliveryappproject/domain/store/service/StoreServiceTest.java
@@ -11,7 +11,7 @@ import com.example.deliveryappproject.domain.store.entity.Store;
 import com.example.deliveryappproject.domain.store.entity.StoreState;
 import com.example.deliveryappproject.domain.store.repository.StoreRepository;
 import com.example.deliveryappproject.domain.user.entity.User;
-import com.example.deliveryappproject.domain.user.entity.UserRole;
+import com.example.deliveryappproject.domain.user.enums.UserRole;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
